+++ src/main/java/cn/czfshine/app/store/repository/ProductRepository.java
@@ -11,5 +11,5 @@ import org.springframework.stereotype.Repository;
 
 
 public interface ProductRepository extends JpaRepository<Product,Integer> {
-    Product findFirstByGan(Long gan);
+    Product findFirstByGan(Integer gan);
 }
